Show time and date on forecast update timestamps
Formatter.TimepointRelative shows plain time without date if the
timestamp is within a week. Thus it's quite confusing if the
update is further than 24h. Inside of 24h period it would be ok, but
then again update when that time is passed would need some extra
handling too.

Think time and date should be fine for separately opened menu and
an error case.