[sailfish-weather] Ensure oberervation fetch updates the correct weather entry. JB#58739
When switching weather locations, the observations would be reloaded.
Invariably the locationId would then be updated before the fetch
completed, so that the wrong locationId entry was updated in the weather
model.

This change keeps a copy of the locationID, so that the correct entry
will be updated when the fetch completes.